+++ internal/provider/data_source_instance.go
@@ -57,6 +57,7 @@ func (d *instanceDataSource) Schema(ctx context.Context, req datasource.SchemaRe
 			},
 			"short_name": schema.StringAttribute{
 				MarkdownDescription: "Instance's short name",
+				Optional:            true,
 				Computed:            true,
 			},
 			"description": schema.StringAttribute{
@@ -82,7 +83,7 @@ func (d *instanceDataSource) Read(ctx context.Context, req datasource.ReadReques
 	if data.ShortName.ValueString() == "" {
 		attributes.Set("name", data.Name.ValueString())
 	} else {
-		attributes.Set("shortName", data.ShortName.ValueString())
+		attributes.Set("short_name", data.ShortName.ValueString())
 	}
 
 	instanceResponse, err := d.provider.client.GetInstancesByAttributes(attributes)