+++ internal/tests/ApplicationDataSource/application_data_source_test.go
@@ -29,6 +29,28 @@ func TestApplicationDataSource(t *testing.T) {
 	})
 }
 
+func TestApplicationDataSourceByShortName(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: GetProviderConfig(t).ProtoV6ProviderFactories,
+		ExternalProviders:        GetProviderConfig(t).ExternalProviders,
+		Steps: []resource.TestStep{
+			{
+				Config: GetProvider() + GetFile("axual_application.tf"),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.axual_application.tf-test-app-imported-by-short-name", "name", "tf-test-app"),
+					resource.TestCheckResourceAttr("data.axual_application.tf-test-app-imported-by-short-name", "application_type", "Custom"),
+					resource.TestCheckResourceAttr("data.axual_application.tf-test-app-imported-by-short-name", "short_name", "tf_test_app_short"),
+					resource.TestCheckResourceAttr("data.axual_application.tf-test-app-imported-by-short-name", "application_id", "tf.test.app"),
+					resource.TestCheckResourceAttrPair("data.axual_application.tf-test-app-imported-by-short-name", "owners", "data.axual_group.test_group", "id"),
+					resource.TestCheckResourceAttr("data.axual_application.tf-test-app-imported-by-short-name", "type", "Java"),
+					resource.TestCheckResourceAttr("data.axual_application.tf-test-app-imported-by-short-name", "visibility", "Public"),
+					resource.TestCheckResourceAttr("data.axual_application.tf-test-app-imported-by-short-name", "description", "Axual's TF Test Application"),
+				),
+			},
+		},
+	})
+}
+
 func TestConnectorApplicationDataSource(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		ProtoV6ProviderFactories: GetProviderConfig(t).ProtoV6ProviderFactories,