+++ internal/tests/EnvironmentDataSource/environment_data_source_test.go
@@ -32,3 +32,24 @@ func TestEnvironmentDataSource(t *testing.T) {
 		},
 	})
 }
+
+func TestEnvironmentDataSourceGetByShortName(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: GetProviderConfig(t).ProtoV6ProviderFactories,
+		ExternalProviders:        GetProviderConfig(t).ExternalProviders,
+		Steps: []resource.TestStep{
+			{
+				Config: GetProvider() + GetFile("axual_environment.tf"),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.axual_environment.tf-test-env-imported-short-name", "name", "tf-development"),
+					resource.TestCheckResourceAttr("data.axual_environment.tf-test-env-imported-short-name", "short_name", "tfdev"),
+					resource.TestCheckResourceAttr("data.axual_environment.tf-test-env-imported-short-name", "description", "This is the terraform testing environment"),
+					resource.TestCheckResourceAttr("data.axual_environment.tf-test-env-imported-short-name", "color", "#19b9be"),
+					resource.TestCheckResourceAttr("data.axual_environment.tf-test-env-imported-short-name", "visibility", "Private"),
+					resource.TestCheckResourceAttr("data.axual_environment.tf-test-env-imported-short-name", "authorization_issuer", "Auto"),
+					resource.TestCheckResourceAttrPair("data.axual_environment.tf-test-env-imported", "owners", "data.axual_group.test_group", "id"),
+				),
+			},
+		},
+	})
+}