+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [master](https://github.com/Axual/terraform-provider-axual/blob/master) - TBR
 * Add User data source
 * Guide for JSON schema
+* Added support for searching by shortName in Instance, Environment, Application Data Sources.
 
 ## [2.5.5](https://github.com/Axual/terraform-provider-axual/releases/tag/v2.5.5) - 2025-03-07
 * Removed unused method for `/groups/{uid}/members/{uid}`